+++ demisto_sdk/validation/secrets.py
@@ -68,15 +68,19 @@ class SecretsValidator():
             secrets_file_paths = self.get_all_diff_text_files(branch_name, is_circle)
             secrets_found = self.search_potential_secrets(secrets_file_paths)
             if secrets_found:
-                secrets_found_string = 'Secrets were found in the following files:\n'
+                secrets_found_string = 'Secrets were found in the following files:'
                 for file_name in secrets_found:
-                    secrets_found_string += ('\nFile Name: ' + file_name)
+                    secrets_found_string += ('\n\nIn File: ' + file_name + '\n')
+                    secrets_found_string += '\nThe following expressions were marked as secrets: \n'
                     secrets_found_string += json.dumps(secrets_found[file_name], indent=4)
+
                 if not is_circle:
-                    secrets_found_string += '\nRemove or whitelist secrets in order to proceed, then re-commit\n'
+                    secrets_found_string += '\n\nRemove or whitelist secrets in order to proceed, then re-commit\n'
+
                 else:
-                    secrets_found_string += 'The secrets were exposed in public repository,' \
+                    secrets_found_string += '\n\nThe secrets were exposed in public repository,' \
                                             ' remove the files asap and report it.\n'
+
                 secrets_found_string += 'For more information about whitelisting visit: ' \
                                         'https://github.com/demisto/internal-content/tree/master/documentation/secrets'
                 print_error(secrets_found_string)